src/tests/use-online-status.test.ts
import { renderHook, act } from '@testing-library/react';
import { useOnlineStatus } from '../hooks/use-online-status';

describe('useOnlineStatus hook', () => {
  it('should return true when online', () => {
    const { result } = renderHook(() => useOnlineStatus());
    expect(result.current).toBe(true);
  });

  it('should return false when offline', () => {
    const { result } = renderHook(() => useOnlineStatus());

    act(() => {
      window.dispatchEvent(new Event('offline'));
    });

    expect(result.current).toBe(false);

    act(() => {
      window.dispatchEvent(new Event('online'));
    });

    expect(result.current).toBe(true);
  });
});
